requirements.txt
flask
aiogram
python-dotenv
requests
psycopg2-binary
python-telegram-bot
